+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added dns lookup in the radio script to find the active servers urls
 - Added opentube library, that is used to search for videos using the YouTube Public Data without YouTubeAPI
+- Added -sv/--search-video argument to PeerTube script
+- Added -x/--extra-info argument to PeerTube script to show additional metadata
+
+### Removed
+- Removed the Wasd.tv script
+
+### Fixed
+- Fixed the dailymotion script
 
 ## [v1.7.8] - 2025-04-19